+++ src/main/java/christmas/domain/Event.java
@@ -1,10 +1,10 @@
 package christmas.domain;
 
-import christmas.domain.contants.Badge;
-import christmas.domain.contants.Gift;
-import christmas.domain.contants.event.EventDiscount;
-import christmas.domain.contants.event.EventValue;
-import christmas.domain.contants.menu.MenuGroup;
+import christmas.domain.constants.Badge;
+import christmas.domain.constants.Gift;
+import christmas.domain.constants.event.EventDiscount;
+import christmas.domain.constants.event.EventValue;
+import christmas.domain.constants.menu.MenuGroup;
 
 import java.util.Arrays;
 import java.util.Collections;
