+++ src/main/java/christmas/domain/constants/Gift.java
@@ -1,4 +1,4 @@
-package christmas.domain.contants;
+package christmas.domain.constants;
 
 import java.util.function.Function;
