+++ src/main/java/christmas/domain/constants/menu/BeverageMenu.java
@@ -1,13 +1,14 @@
-package christmas.domain.contants.menu;
+package christmas.domain.constants.menu;
 
-public enum DessertMenu implements MenuInterface {
-    CHOCO_CAKE("초코케이크", 15_000),
-    ICE_CREAM("아이스크림", 5_000);
+public enum BeverageMenu implements MenuInterface {
+    ZERO_COLA("제로콜라", 3_000),
+    RED_WINE("레드와인", 60_000),
+    CHAMPAGNE("샴페인", 25_000);
 
     private final String name;
     private final int price;
 
-    DessertMenu(String name, int price) {
+    BeverageMenu(String name, int price) {
         this.name = name;
         this.price = price;
     }
