+++ src/main/java/christmas/domain/constants/menu/MainMenu.java
@@ -1,14 +1,15 @@
-package christmas.domain.contants.menu;
+package christmas.domain.constants.menu;
 
-public enum AppetizerMenu implements MenuInterface {
-    MUSHROOM_CREAM_SOUP("양송이수프", 6_000),
-    TAPAS("타파스", 5_500),
-    CAESAR_SALAD("시저샐러드", 8_000);
+public enum MainMenu implements MenuInterface {
+    T_BONE_STEAK("티본스테이크", 55_000),
+    BBQ_RIBS("바비큐립", 54_000),
+    SEAFOOD_PASTA("해산물파스타", 35_000),
+    CHRISTMAS_PASTA("크리스마스파스타", 25_000);
 
     private final String name;
     private final int price;
 
-    AppetizerMenu(String name, int price) {
+    MainMenu(String name, int price) {
         this.name = name;
         this.price = price;
     }
