+++ src/main/java/christmas/view/constants/InputMessage.java
@@ -1,4 +1,4 @@
-package christmas.view.contants;
+package christmas.view.constants;
 
 public enum InputMessage {
     VISIT_DATE("12월 중 식당 예상 방문 날짜는 언제인가요? (숫자만 입력해 주세요!)"),
